ARM: dts: dra7: fix control module core instance nodes
The old names for these nodes were misleading, and also the address spaces
for these didn't match the hardware. Fixed by renaming the nodes and
fixing the memory maps to reflect actual hardware implementation. Existing
users for the syscons updated accordingly. Support for the new dra7
control module compatible strings is going to be added in a separate patch.

Signed-off-by: Tero Kristo <t-kristo@ti.com>