ARM: OMAP2+: hwmod: add support for rerouting hwmod links
If a hwmod with same name is already registered, reroute the links to
use the existing one rather than fail. This is needed when hwmod data
is made into a separate module, and for example, l3 bus is already
registered under different hwmod. The late module init will use
an l3 dummy hwmod instead, which will be rerouted to use the real
l3 hwmod.

Signed-off-by: Tero Kristo <t-kristo@ti.com>