ARM: OMAP2+: PRM: add API for providing pdata for the PRCM drivers
This is required for passing some platform functions to the driver.
Initially, we only pass pcs_legacy_init function, which is needed
for setting up the pinctrl interrupt properly, but the pdata struct
can be expanded later if needed.

Also, removed some mach-omap2 specific headers which are no longer
needed for anything.

This patch is done in preparation of making PRCM a separate driver.

Signed-off-by: Tero Kristo <t-kristo@ti.com>