ARM: OMAP4: powerdomain: remove obsolete check for missing partition
cpu_is_* calls should be avoided in the driver code, and the easiest
way to get rid of this in this case is just to drop the redundant check
completely. The check has only ever been used on OMAP4, and OMAP5 and
DRA7xx have managed well without it, so there is little need to
complicate the code with this check.

Signed-off-by: Tero Kristo <t-kristo@ti.com>