ARM: OMAP2+: clockdomain: remove cpu type checks
Most of the omap2 / omap3 hwmod data already has clockdomains defined,
so the soc type check here is redundant. Also, the hwmod code itself
checks against missing clockdomain entry, and will not call the
functions in the first place. Thirdly, soc_is_* checks should be avoided
in the driver code, and we are in the process of migrating the
clockdomain code under the PRCM driver. Thus, the cpu_is_* checks
are removed.

Remove a couple of redundant header files while we are touching the
file.

Signed-off-by: Tero Kristo <t-kristo@ti.com>