ARM: OMAP2+: hwmod: assign hwmod-ck as main clock from DT if available
Fix the main clock assignment to assign hwmod-ck from DT as the main clock
if available. If main clock is assigned via DT, the direct PRCM access
for module handling is not used on OMAP4+ architectures either, as it is
assumed the main clock will be doing this instead.

This patch also drops the obsolete "_mod_ck" search as the implementation
required for this was not accepted upstream.

Signed-off-by: Tero Kristo <t-kristo@ti.com>