remoteproc: add a rproc_set_firmware() API
A new API, rproc_set_firmware() is added to allow the remoteproc platform
drivers to be able to configure a custom firmware name that is different
from the default name used during remoteproc registration. This function
is being introduced to provide a kernel-level equivalent of the current
sysfs interface but limited to remoteproc drivers. This allows some
remoteproc-drivers to choose different firmwares at runtime when the
remote processor is not running based on the functional feature it is
providing using that remote processor.

Signed-off-by: Suman Anna <s-anna@ti.com>