bus: ti-sysc: add support for controlling the IP directly from driver
IPs like IOMMUs and remoteprocs have strict sequencing logic for powering
them up and down. To allow more granular control for these for the device
drivers, extend the ti-sysc interface with new APIs for direct control.
Also, add new flag for this feature so that it can be limited to the
specific IPs only.

Signed-off-by: Tero Kristo <t-kristo@ti.com>